+++ vulkano/src/pipeline/graphics_pipeline/builder.rs
@@ -42,6 +42,7 @@ use pipeline::shader::GraphicsEntryPointAbstract;
 use pipeline::shader::GraphicsShaderType;
 use pipeline::shader::ShaderInterfaceDefMatch;
 use pipeline::shader::SpecializationConstants;
+use pipeline::vertex::BufferlessDefinition;
 use pipeline::vertex::SingleBufferDefinition;
 use pipeline::vertex::VertexDefinition;
 use pipeline::viewport::Scissor;
@@ -83,7 +84,7 @@ struct TessInfo<Tcs, Tcss, Tes, Tess> {
 }
 
 impl
-    GraphicsPipelineBuilder<SingleBufferDefinition<()>,
+    GraphicsPipelineBuilder<BufferlessDefinition,
                             EmptyEntryPointDummy,
                             (),
                             EmptyEntryPointDummy,
@@ -99,7 +100,7 @@ impl
     pub(super) fn new() -> Self {
         unsafe {
             GraphicsPipelineBuilder {
-                vertex_input: SingleBufferDefinition::new(), // TODO: should be empty attrs instead
+                vertex_input: BufferlessDefinition,
                 vertex_shader: None,
                 input_assembly: vk::PipelineInputAssemblyStateCreateInfo {
                     sType: vk::STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO,